Implement impl in the service

+++ src/main/java/org/nikisurance/repository/DependentRepository.java
@@ -3,5 +3,5 @@ package org.nikisurance.repository;
 import org.nikisurance.entity.Dependent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DependentRepository extends JpaRepository<Dependent, Integer> {
+public interface DependentRepository extends JpaRepository<Dependent, Long> {
 }
